backend/requirements.txt
fastapi
uvicorn
pandas
pydantic
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
